Reinstate ss-broken styling in editor.css
Already in place for framework/admin/css/editor.css, but the theme file overwrites this.
Shows red background styling if links are broken in CMS editor, which is an essential hint for CMS authors.
At the moment, they'll review the "Pages with broken links" report, but have no clue which of the links is broken.